Publishing for dlorenc!

+++ scripts/step-5.sh
@@ -30,8 +30,10 @@ git status
 # Sign the root and targets
 ./tuf publish -repository $REPO
 # Clear and copy into the repository/
-rm -r repository/ if [ -f $REPO/repository/1.root.json ]; then
-cp -r $REPO/repository/ repository/
+rm -r repository/
+if [ -f $REPO/repository/1.root.json ]; then
+    cp -r $REPO/repository/ repository/
+fi
 
 if [ -n "$NO_PUSH" ]; then
     echo "Skipping push, exiting early..."
